Moved ExtensionFunctions to about directory because they were only being used there. Added comments.
