+++ androidutilities/src/main/java/com/heyzeusv/androidutilities/compose/about/ExtensionFunctions.kt
@@ -1,11 +1,19 @@
-package com.heyzeusv.androidutilities.compose.util
+package com.heyzeusv.androidutilities.compose.about
 
 import androidx.core.text.HtmlCompat
 
+/**
+ *  A version of [String.ifBlank] that works on nullable [String]. Returns this [String] if it is
+ *  not null, not empty, and doesn't consist solely of whitespace characters, or [defaultValue]
+ *  otherwise.
+ */
 fun String?.ifNullOrBlank(defaultValue: String): String {
     return if (this.isNullOrBlank()) defaultValue else this
 }
 
+/**
+ *  Formats String by removing new line characters that are in the middle of sentences using Regex.
+ */
 fun String.formatContent(): String {
     val removeSpacing = HtmlCompat.fromHtml(
         this.replace("\n", "<br/>"),
